C++ net: f32 flat weights (templated NetT<T>) for SIMD; gradient check kept in f64
Refactor net.h to NetT<T> with FLAT contiguous per-layer weights (row-major
dout*din), blocked forward + backprop. Net = NetT<float> for production (wide
SIMD); the gradient-check test uses NetT<double> so it stays rigorous (same code
path). Cascade: encode() and Sample x/pi are now float; net_io stores weights
flat per layer (fresh-start format); selfplay/mcts/buffer unchanged in behavior.

All 5 ctests pass (gradient check, scoring, game sanity, antithetic, netio,
orchestration; pi-sum tolerance relaxed to float epsilon). cz_pc --dry self-play
2.16 -> 7.5 g/s cumulatively (~3.5x from the naive baseline).

Honest note: f32 alone added only ~10% over blocked-f64 — the remaining cost has
shifted to ALLOCATION (forward builds an acts[] it doesn't need for inference;
CribGame clones allocate ~10 small vectors per MCTS sim). Next perf levers: an
inference-only forward (no acts, scratch buffers) and fixed-size game arrays.
Bumped to 0.1.35-dev.6.

+++ core/game.h
@@ -160,8 +160,8 @@ private:
 
 public:
   // ----- net interface: fixed-length features from `player`'s information set (opponent cards hidden) -----
-  std::vector<double> encode(int player) const {
-    std::vector<double> fv;
+  std::vector<float> encode(int player) const {
+    std::vector<float> fv;
     fv.reserve(INPUT_DIM);
     int opp = 1 - player;
     auto pushCard = [&](const Card* c) {
