+++ core/sample.h
@@ -6,8 +6,8 @@
 namespace cz {
 
 struct Sample {
-  std::vector<double> x;     // encoded state (INPUT_DIM)
-  std::vector<double> pi;    // MCTS visit distribution over slots (NPOL)
+  std::vector<float> x;      // encoded state (INPUT_DIM)
+  std::vector<float> pi;     // MCTS visit distribution over slots (NPOL)
   std::vector<bool> legal;   // legal slot mask (NPOL)
   double z;                  // game outcome from the acting player's view (+1 win / -1 loss)
 };
